feat(pages): Introduce react-router and split into pages
